[#87] refactor: 위젯을 다시 분리해 재작업 시작
