+++ lib/features/effects/balloon_animation/balloon_manager.dart
@@ -29,20 +29,6 @@ class BalloonManager extends StateNotifier<Map<Key, BalloonWidget>> {
 
   void addBalloon({required String imageUrl}) {
     final balloonWidgetKey = UniqueKey();
-
-    // state.addEntries(
-    //   <Key, BalloonWidget>{
-    //     balloonWidgetKey: BalloonWidget(
-    //       key: balloonWidgetKey,
-    //       imageUrl: imageUrl,
-    //       notifyWidgetIsDisposed: (Key widgetKey) {
-    //         state = Map.of(state..remove(widgetKey));
-    //       },
-    //       onTapWithOffset: onTapCallbackWithTappedPositionOffset ?? (_) {},
-    //       radius: radius,
-    //     ),
-    //   }.entries,
-    // );
     final newBalloon = {
       balloonWidgetKey: BalloonWidget(
         key: balloonWidgetKey,
@@ -55,10 +41,10 @@ class BalloonManager extends StateNotifier<Map<Key, BalloonWidget>> {
       ),
     };
 
-    // state = {
-    //   ...state,
-    //   ...newBalloon,
-    // };
+    state = {
+      ...state,
+      ...newBalloon,
+    };
   }
 }
 
@@ -147,14 +133,12 @@ class _BalloonWidgetState extends State<BalloonWidget>
 
   void notifyWidgetIsDisposed() {
     _animationController.dispose();
-    print("TAP2");
     widget.onTapWithOffset(
       Offset(
         balloonModel.x + 50 * sin(_animation.value) + 65, // radius of photo
         balloonModel.y + 25 * cos(_animation.value) + 65, // radius of photo
       ),
     );
-    print("TAP");
     widget.notifyWidgetIsDisposed(widget.key);
   }
 }
