+++ lib/features/effects/effects.dart
@@ -1,3 +1,5 @@
+export 'animation_sample_page.dart';
 export 'balloon_animation/balloon_manager.dart';
-export 'balloon_sample_page.dart';
 export 'emoji_firework_animation/emoji_firework_manager.dart';
+export 'text_animation/text_bubble_manager.dart';
+export 'text_animation/text_bubble_widget.dart';
